yaR-server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altgraph==0.17.3
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
boto3==1.34.35
botocore==1.34.35
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
clarifai==9.8.2
clarifai-grpc==9.8.6
cohere==4.27
colorama==0.4.6
comm==0.1.4
contextlib2==21.6.0
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.6.1
debugpy==1.8.0
decorator==4.4.2
dill==0.3.7
distro==1.9.0
Django==4.2.10
django-cors-headers==4.3.1
dnspython==2.4.2
easyocr==1.7.1
exceptiongroup==1.1.3
executing==1.2.0
faiss-cpu==1.7.4
fastavro==1.8.2
ffprobe==0.5
filelock==3.12.4
fonttools==4.42.1
fpdf==1.7.2
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.3
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.0
google-cloud==0.34.0
google-cloud-speech==2.24.1
google-cloud-texttospeech==2.16.1
google-cloud-vision==3.7.0
google-generativeai==0.3.2
googleapis-common-protos==1.60.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
halo==0.0.31
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.17.3
idna==3.4
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.0.1
imutils==0.5.4
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
langchain==0.0.306
langsmith==0.0.41
lazy_loader==0.3
log-symbols==0.0.14
macholib @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
manifest-ml==0.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
ninja==1.11.1.1
nlpcloud==1.1.44
numexpr==2.8.7
numpy==1.25.2
openai==1.8.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openlm==0.0.5
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
playsound==1.3.0
plotly==5.17.0
proglog==0.1.10
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.24.3
psutil==5.9.5
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pvcheetah==2.0.0
pvcheetahdemo==2.0.0
pvrecorder==1.2.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
pygame==2.5.2
Pygments==2.16.1
pymongo==4.5.0
pyparsing==3.0.9
pypdf==3.16.2
PyPDF2==3.0.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-rapidjson==1.11
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
redis==5.0.1
regex==2023.8.8
reportlab==4.0.9
requests==2.31.0
rich==13.4.2
rsa==4.9
s3transfer==0.10.0
safetensors==0.3.3
schema==0.7.5
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.12.2
shapely==2.0.3
six @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.0
sounddevice==0.4.6
spinners==0.0.24
SQLAlchemy==2.0.21
sqlitedict==2.1.0
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2024.2.12
tiktoken==0.5.1
tokenizers==0.13.3
torch==2.2.0
torchvision==0.17.0
tornado==6.3.3
tqdm==4.64.1
traitlets==5.10.0
transformers==4.33.3
tritonclient==2.34.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
wcwidth==0.2.6
webrtcvad==2.0.10
yarl==1.9.2
zipp==3.16.2
